user on rating

+++ catalogo_filmes/lib/controller/firebaseController.dart
@@ -6,21 +6,39 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
 
 import '../models/movie.dart';
+import '../models/user.dart';
+import '../services/auth_service.dart';
 
 class FirebaseController {
   FirebaseController();
   final _baseUrl =
       dotenv.get('FIREBASE_URL'); // Colocar valor da variável no .env
 
+  final user = AuthService().auth.currentUser!;
+
   Future<void> saveRatingForm(Map<String, Object> data) {
     bool hasId = data['id'] != null;
     bool hasComment = data['comment'] != null;
 
+    String userName =
+        user.displayName != null ? user.displayName.toString() : 'Sem nome';
+    String photoURL = user.photoURL != null
+        ? user.photoURL.toString()
+        : '../assets/image/person-icon.png';
+
+    UserApp userLocal = UserApp(
+      uid: user.uid.toString(),
+      userName: userName,
+      profilePictureUrl: photoURL,
+    );
+    print(user);
+    print("[===DATA]: ${userLocal.profilePictureUrl}");
     final rating = Rating(
       id: hasId ? data['id'] as String : Random().nextDouble().toString(),
       movie: data['movie'] as Movie,
       value: data['value'] as double,
       comment: hasComment ? data['comment'] as String : null,
+      userApp: userLocal,
     );
 
     if (hasId) {
@@ -36,6 +54,11 @@ class FirebaseController {
           "movie": rating.movie,
           "value": rating.value,
           "comment": rating.comment,
+          "userApp": {
+            "uid": rating.userApp.uid,
+            "userName": rating.userApp.userName,
+            "profilePictureUrl": rating.userApp.profilePictureUrl,
+          },
         }));
     return future.then((response) {
       final id = jsonDecode(response.body)['name'];
@@ -50,7 +73,7 @@ class FirebaseController {
           Uri.parse(url),
           body: jsonEncode(
             {
-              "movie": rating.movie,
+              //"movie": rating.movie,
               "value": rating.value,
               "comment": rating.comment,
             },
@@ -60,15 +83,17 @@ class FirebaseController {
             "Status: ${response.statusCode.toString()} = ${response.reasonPhrase.toString()}"));
   }
 
-  Future<void> deleteRatingInFirebase(String id) {
-    final url = "$_baseUrl/ratings_and_reviews/$id.json";
-
-    return http.delete(Uri.parse(url)).then((response) {
-      print(response.statusCode);
-      if (response.statusCode >= 400) {
-        throw Exception("Could not delete!");
-      }
-    });
+  Future<void> deleteRatingInFirebase(Rating rating) async {
+    final url = "$_baseUrl/ratings_and_reviews/${rating.id}.json";
+    bool isOwner = rating.userApp.uid == user.uid ? true : false;
+    if (isOwner) {
+      return http.delete(Uri.parse(url)).then((response) {
+        print(response.statusCode);
+        if (response.statusCode >= 400) {
+          throw Exception("Could not delete!");
+        }
+      });
+    }
   }
 
   Future<Rating> getRatingInFirebaseById(String id) async {
